% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{abilities}
\alias{abilities}
\title{Abilities}
\format{
A tibble with 2 variables:
\describe{
\item{name}{Ability name}
\item{attribs}{Ability attributes, formatted as a nested list}
}
}
\source{
\url{https://docs.opendota.com/#tag/constants}
}
\usage{
abilities
}
\description{
Hero abilities in Dota 2
}
\keyword{datasets}
