% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{ability_ids}
\alias{ability_ids}
\title{Ability Ids}
\format{
A tibble with 2 variables:
\describe{
\item{ability_id}{Ability id number}
\item{ability_name}{Ability name}
}
}
\source{
\url{https://docs.opendota.com/#tag/constants}
}
\usage{
ability_ids
}
\description{
Mapping of ability names to ids
}
\keyword{datasets}
