\name{antitable}
\alias{antitable}
\docType{data}
\title{
Data for disadvantages.
}
\description{
Data for disadvantages.
}
\usage{data("antitable")}
\format{
  The format is:
 chr [1:109, 1:109] "0" "0.0249" "0.0072" "0.0601" "0.0262" "0.0526" "-2e-04" ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:109] "antimage" "axe" "bane" "bloodseeker" ...
  ..$ : NULL
}
\keyword{datasets}
