% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match.R
\name{get_match}
\alias{get_match}
\title{Get Match Data}
\usage{
get_match(match_id, as_tibble = TRUE)
}
\arguments{
\item{match_id}{Match ID to query}

\item{as_tibble}{logical. If TRUE, will return as a 1-row tibble with some list-cols. Otherwise, will return as a nested list.}
}
\value{

}
\description{
This function queries data for a given match. It will likely be split into other functions that return more specific information
}
\examples{
\dontrun{
get_match('6019587919')
}
}
