% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parsed_matches.R
\name{get_parsed_matches}
\alias{get_parsed_matches}
\title{Get Parsed Matches}
\usage{
get_parsed_matches(less_than_match_id = NULL)
}
\arguments{
\item{less_than_match_id}{Optional. If set, will return matches with a match id lower than this value.}
}
\value{

}
\description{
Get a vector of random match ids for parsed matches.
}
\examples{
\dontrun{
#will return a random vector of parsed matches
a <- get_parsed_matches()

#will return a random vector of parsed matches with a match id less than 6192935701
b <- get_parsed_matches(less_than_match_id = 6192935701)
}
}
