% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_account.R
\name{get_player_account_info}
\alias{get_player_account_info}
\title{Get Player Account Information}
\usage{
get_player_account_info(player_id, tidy = TRUE)
}
\arguments{
\item{player_id}{Steam32 account id}
}
\value{

}
\description{
Get Player Account Information
}
\examples{
\dontrun{
get_player_account_info('108887322')
}
}
