% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_recent_matches.R
\name{get_player_recent_matches}
\alias{get_player_recent_matches}
\title{Get Player's Recent Matches}
\usage{
get_player_recent_matches(player_id, as_tibble = TRUE)
}
\arguments{
\item{player_id}{Steam32 account id}

\item{as_tibble}{logical. If TRUE, will return the result as a tibble where each row is a match. If false, as an 'rdota' object with the .}
}
\value{

}
\description{
Get Player's Recent Matches
}
\examples{
\dontrun{
get_player_recent_matches('108887322')
}
}
