% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro_matches.R
\name{get_pro_matches}
\alias{get_pro_matches}
\title{Get Pro Matches}
\usage{
get_pro_matches(less_than_match_id = NULL)
}
\arguments{
\item{less_than_match_id}{Optional. Get matches with a match ID lower than this value}
}
\value{
a tibble containing a limited set of information about the match. Passing match ids to \code{\link{get_match}}
can provide more data about individual matches
}
\description{
Get a selection of randomly sampled pro matches.
}
\examples{
\dontrun{
a <- get_pro_matches()

#or if you want to limit to less than a certain match id
b <- get_pro_matches(less_than_match_id = 6257294935)
}
}
