% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_public_matches.R
\name{get_public_matches}
\alias{get_public_matches}
\title{Get Public Matches}
\usage{
get_public_matches(
  less_than_match_id = NULL,
  mmr_ascending = NULL,
  mmr_descending = NULL,
  use_hero_names = TRUE
)
}
\arguments{
\item{less_than_match_id}{Optional. Get matches with a match ID lower than this value}

\item{mmr_ascending}{Optional. If TRUE, will order results by ascending MMR. Cannot be TRUE if mmr_descending is TRUE.}

\item{mmr_descending}{Optional. If TRUE, will order results by descending MMR. Cannot be TRUE if mmr_ascending is TRUE.}

\item{use_hero_names}{Logical. If TRUE, will return hero names instead of IDs. If FALSE, will return hero IDs.}
}
\value{
a tibble containing a limited set of information about the match
}
\description{
Get a selection of randomly sampled public matches.
}
\examples{
\dontrun{
a <- get_public_matches()
b <- get_public_matches(less_than_match_id = 6192935701)
}
}
