% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Core Get Response}
\usage{
get_response(resource, ...)
}
\arguments{
\item{resource}{name of the resource to pass to the OpenDota API}

\item{...}{query arguments}
}
\value{

}
\description{
This function powers the API requests in rdota. It's for internal purposes and shouldn't be called directly by the user.
}
