\name{globalonline}
\alias{globalonline}
\title{
Function showing the amount of people who is currently in game
}
\description{
Function showing the amount of people who is currently in game
}
\usage{
globalonline()
}

\details{
The data comes from \url{dotamax.com}. The data come from Steamapi, but I avoid to use Steamapi if the data could be found elsewhere.
}
\value{
a numeric element that indicates the amount of people who is currently in game.
}

\author{
Xiao Lei. Send me email if you have any great idea or you are willing to help.
\email{leixiao322@hotmail.com}
}
\examples{
globalonline()
}
