% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{heroes}
\alias{heroes}
\title{Heroes}
\format{
A tibble with 29 variables:
\describe{
\item{id}{Hero id number}
\item{name}{'Proper' hero name}
\item{localized_name}{'Common' hero name}
\item{primary_attr}{Primary attribute}
\item{attack_type}{Attack type (melee or ranged)}
\item{roles}{Recommended/common roles for hero}
\item{img}{Hero image url. Prefix with media.steampowered.com}
\item{icon}{Hero icon url. Prefix with media.steampowered.com}
\item{base_health}{Hero base health}
\item{base_health_regen}{Hero base health regeneration}
\item{base_mana}{Hero base mana}
\item{base_mana_regen}{Hero base mana regeneration}
\item{base_armor}{Hero base armor}
\item{base_mr}{Hero base magic resistance}
\item{base_attack_min}{Hero base attack damage minimum}
\item{base_attack_max}{Hero base attack damage maximimum}
\item{base_str}{Hero base strength}
\item{base_agi}{Hero base agility}
\item{base_int}{Hero base intelligence}
\item{str_gain}{Strength gained per level}
\item{agi_gain}{Agility gained per level}
\item{int_gain}{Intelligence gained per level}
\item{attack_range}{Hero attack range}
\item{projectile_speed}{Travel speed of the hero attack projectile}
\item{attack_rate}{Base time between hero attacks}
\item{move_speed}{Base hero movement speed}
\item{turn_rate}{Base speed at which hero turns in place}
\item{cm_enabled}{Logical. Is the hero enabled in captain's mode?}
\item{legs}{Number of legs the hero has}
}
}
\source{
\url{https://docs.opendota.com/#tag/constants}
}
\usage{
heroes
}
\description{
Information on playable heroes in Dota 2.
}
\keyword{datasets}
