% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{item_ids}
\alias{item_ids}
\title{Item Ids}
\format{
A tibble with 2 variables:
\describe{
\item{item_id}{Item id number}
\item{item_name}{Item name}
}
}
\source{
\url{https://docs.opendota.com/#tag/constants}
}
\usage{
item_ids
}
\description{
Mapping of item names to ids
}
\keyword{datasets}
