% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{items}
\alias{items}
\title{Items}
\format{
A tibble with 14 variables:
\describe{
\item{hint}{Text hint(s) displayed on item tooltip}
\item{id}{Item id number}
\item{img}{Item image url. Prefix with media.steampowered.com}
\item{dname}{Item name}
\item{qual}{Item quality}
\item{cost}{Item cost, in gold}
\item{notes}{Additional item notes}
\item{attrib}{Item attributes, as a nested list}
\item{mc}{Mana cost to use item}
\item{cd}{Item cooldown, in seconds}
\item{lore}{Item lore}
\item{components}{Components used to create item, if applicable}
\item{created}{Logical. Is the item created from other items?}
\item{charges}{Number of item charges available up on purchase/creation, if applicable}
}
}
\source{
\url{https://docs.opendota.com/#tag/constants}
}
\usage{
items
}
\description{
Available items in Dota 2
}
\keyword{datasets}
