\name{matchtable}
\alias{matchtable}
\docType{data}
\title{
Data for advantages.
}
\description{
Data for advantages
}
\usage{data("matchtable")}
\format{
  The format is:
 chr [1:109, 1:109] "0" "0.0093" "0.0325" "-0.0198" "0.0046" "-0.0756" "0.0153" ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:109] "antimage" "axe" "bane" "bloodseeker" ...
  ..$ : NULL
}
\keyword{datasets}
