% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_parsed_match.R
\name{print.parsed_match}
\alias{print.parsed_match}
\title{print method for parsed_match class}
\usage{
\method{print}{parsed_match}(x, ...)
}
\arguments{
\item{x}{parsed_match object}

\item{...}{currently unused}
}
\value{
Prints summary information for a parsed_match object
}
\description{
print method for parsed_match class
}
