% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rdota_match.R
\name{print.rdota_match}
\alias{print.rdota_match}
\title{print method for rdota_match class}
\usage{
\method{print}{rdota_match}(x, ...)
}
\arguments{
\item{x}{rdota_match object}

\item{...}{currently unused}
}
\value{
Prints the content element of an rdota_match object
}
\description{
print method for rdota_match class
}
