% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_lane_roles.R
\name{pull_lane_roles}
\alias{pull_lane_roles}
\title{Pull Lanes and Roles}
\usage{
pull_lane_roles(obj)
}
\arguments{
\item{obj}{A 'parsed_match' object}
}
\value{
a 10-row tibble containing match/player identifiers, lane played in, and player role
}
\description{
Pull the lanes players played in during the laning phase as well as their roles (i.e. support/core). Roles are estimated using players' lanes and net worth at the 10 minute mark. If a game lasts less than 10 minutes, roles will not be returned.
}
\examples{
\dontrun{
a <- get_match('6387933253')
b <- pull_lane_roles(a)
}
}
