% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_ability_upgrades.R
\name{pull_player_ability_upgrades}
\alias{pull_player_ability_upgrades}
\title{Pull Player Ability Upgrades}
\usage{
pull_player_ability_upgrades(obj)
}
\arguments{
\item{obj}{A 'parsed_match' object}
}
\value{

}
\description{
Pull the ability upgrades chosen for each hero at each level in a match
}
\examples{
\dontrun{
#note that this will only work with a parsed match
a <- get_match('6183712050')
b <- pull_player_ability_upgrades(a)
}
}
