% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_damage_inflicted.R
\name{pull_player_damage_inflicted}
\alias{pull_player_damage_inflicted}
\title{Pull Player Damage Inflicted}
\usage{
pull_player_damage_inflicted(obj, by_hero_recipient = FALSE)
}
\arguments{
\item{obj}{A 'parsed_match' object}

\item{by_hero_recipient}{logical. If FALSE (the default), return the sum of damage by ability to all enemy heroes. If TRUE, return damage done by ability to each enemy hero.}
}
\value{
A tibble, including identifiers and the amount of damage done by each source
}
\description{
Pull the amount of damage inflicted by player by source (abilities, items, and auto-attacks) to all enemy heroes. This will not include damage done to NPCs. With 'by_hero_recipient' set to FALSE (the default), this will return the sum of damage done (by player by source) to all enemy heroes. Setting 'by_hero_recipient' to TRUE will break out damage done by each source to each enemy hero.
}
\examples{
\dontrun{
#note that this requires a parsed match
a <- get_match('6183712050')
b <- pull_player_damage_inflicted(a)

#if you want damage done to each enemy hero by ability, set 'by_hero_recipient' to TRUE:
x <- pull_player_damage_inflicted(a, by_hero_recipient = TRUE)
}
}
