% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_heroes.R
\name{pull_player_heroes}
\alias{pull_player_heroes}
\title{Pull Player Heroes}
\usage{
pull_player_heroes(obj, include_hero_names = TRUE)
}
\arguments{
\item{obj}{An 'rdota_match' object.}

\item{include_hero_names}{logical. If TRUE, will include a column with hero names as well as hero id. If FALSE, will only include hero id.}
}
\value{
A 10x5 tibble
}
\description{
Pull the hero ids/hero names for all players in a given match.
}
\examples{
\dontrun{
a <- get_match('6183712050')
b <- pull_player_heroes(a)
}
}
