% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_kda.R
\name{pull_player_kda}
\alias{pull_player_kda}
\title{Pull Player KDA}
\usage{
pull_player_kda(obj)
}
\arguments{
\item{obj}{A 'parsed_match' object.}
}
\value{

}
\description{
Pull the final kills-deaths-assists (KDA) scores for all players in a match.
}
\examples{
\dontrun{
a <- get_match('6156757097')
b <- pull_player_kda(a)
}
}
