% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_purchase_log.R
\name{pull_player_purchase_logs}
\alias{pull_player_purchase_logs}
\title{Pull Player Purchase Logs}
\usage{
pull_player_purchase_logs(obj)
}
\arguments{
\item{obj}{An 'rdota_match' object.}
}
\value{

}
\description{
Pull the purchase logs, purchased items and time of purchase, for all players in a match
}
\examples{
\dontrun{
#first get an rdota_match object
a <- get_match(6277359804)

#then pull the purchase logs
plog <- pull_player_purchase_logs(a)
}
}
