% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_time_varying.R
\name{pull_player_deny_progression}
\alias{pull_player_deny_progression}
\alias{pull_player_last_hit_progression}
\alias{pull_player_xp_progression}
\alias{pull_player_gold_progression}
\title{Pull Time Varying Player Variables}
\usage{
pull_player_deny_progression(obj)

pull_player_last_hit_progression(obj)

pull_player_xp_progression(obj)

pull_player_gold_progression(obj)
}
\arguments{
\item{obj}{A 'parsed_match' object.}
}
\value{
A 5-column tibble containing match and player identifying information, timestamps (in seconds),
and the value of the given variable at that timestamp
}
\description{
These functions will pull the progression of the given
variable (denies, last hits, experience, or gold) over time for each player in a match.
}
\examples{
\dontrun{
a <- get_match('6156757097')
b <- pull_player_last_hit_progression(a)
}
}
