% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_xy.R
\name{pull_player_xy}
\alias{pull_player_xy}
\title{Pull Player Locations}
\usage{
pull_player_xy(obj)
}
\arguments{
\item{obj}{A 'parsed_match'}
}
\value{
A tibble including identifiers, x/y coordinates, and a count of player appearances at those coordinates
}
\description{
Pull the x/y locations of each player throughout the first 11 minutes of the match.
}
\examples{
\dontrun{
#note that this requires a parsed match
a <- get_match('6183712050')
b <- pull_player_xy(a)
}
}
