% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_time_advantages.R
\name{pull_radiant_gold_adv}
\alias{pull_radiant_gold_adv}
\alias{pull_radiant_xp_adv}
\title{Pull Radiant Gold & XP Advantage}
\usage{
pull_radiant_gold_adv(obj)

pull_radiant_xp_adv(obj)
}
\arguments{
\item{obj}{A 'parsed_match' object}
}
\value{
A 3-column tibble containing the match id, timestamps (in seconds),
and the value of the given variable at that timestamp -- either the radiant gold or experience advantage. Negative
values indicate dire advantage on the given metric
}
\description{
Pull Radiant Gold & XP Advantage
}
\examples{
\dontrun{
a <- get_match('6156757097')
b <- pull_radiant_gold_adv(a)
}
}
