% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_steam32_id.R
\name{retrieve_steam32_id}
\alias{retrieve_steam32_id}
\title{Retrieve 32bit Steam ID}
\usage{
retrieve_steam32_id(long_id)
}
\arguments{
\item{long_id}{The 64-bit Steam ID of the player. Must be passed in as a string}
}
\value{

}
\description{
The OpenDota API requires a 32-bit Steam ID for player queries. This helper function will retrieve a 32-bit ID using the website steamidfinder.com
}
\examples{
\dontrun{
retrieve_steam32_id("76561198069153050")
}
}
