\name{showwinrate}
\alias{showwinrate}
\title{
Show the winrates of all heros
}
\description{
Show the unconditional winrates of all heros.
}
\usage{
showwinrate(language = "English")
}

\arguments{
  \item{language}{
Language for showing the table. There are two options, "English" or "Chinese".
}
}
\details{
The data is from dotamax.com. 
}

\author{
Xiao Lei
}


\examples{
showwinrate()
showwinrate("Chinese")
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
