% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_match.R
\name{tidy_rdota_match}
\alias{tidy_rdota_match}
\title{Tidy an rdota_match object}
\usage{
tidy_rdota_match(match_obj)
}
\arguments{
\item{match_obj}{An rdota_match object.}
}
\value{

}
\description{
Take an rdota_match object -- the result of get_match() -- and turn it into a tibble. Some columns will still contain (deeply) nested lists. Please see pull_*() helper functions to extract data from these columns
}
\examples{
\dontrun{
a <- get_match('6019587919')
tidy_rdota_match(a)
}
}
