\name{winrate}
\alias{winrate}
\docType{data}
\title{
Unconditional winrates of all heros.
}
\description{
Unconditional winrates of all heros.
}
\usage{data("winrate")}
\format{
  A data frame with 109 observations on the following 4 variables.
  \describe{
    \item{\code{Chinese}}{a character vector}
    \item{\code{English}}{a character vector}
    \item{\code{Korean}}{a character vector}
    \item{\code{rate}}{a numeric vector}
  }
}
\examples{
data(winrate)
}
\keyword{datasets}
