% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match.R
\name{get_match}
\alias{get_match}
\title{Get Match Data}
\usage{
get_match(match_id, tidy = TRUE)
}
\arguments{
\item{match_id}{Match ID to query}

\item{tidy}{logical. If TRUE (the default), will return a tidied tibble with class "rdota_match." If FALSE, will return the raw response body as a list.}
}
\value{

}
\description{
This function queries data for a given match. You can use it in conjunction with several pull_*() helpers to get and tidy match data
}
\examples{
\dontrun{
a <- get_match('6019587919')
}
}
