% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_account.R
\name{get_player_account_info}
\alias{get_player_account_info}
\title{Get Player Account Information}
\usage{
get_player_account_info(player_id, content_only = TRUE)
}
\arguments{
\item{player_id}{Steam32 account id}

\item{content_only}{logical. If TRUE, return only the content (i.e. the player account info) of the query. If FALSE, return an 'rdota' object with additional metadata}
}
\value{

}
\description{
Get Player Account Information
}
\examples{
\dontrun{
get_player_account_info('108887322')
}
}
