% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_player_final_items.R
\name{pull_player_final_items}
\alias{pull_player_final_items}
\title{Pull Player Final Items}
\usage{
pull_player_final_items(obj)
}
\arguments{
\item{obj}{A 'parsed_match'  object.}
}
\value{
A tibble with 110 rows and 5 columns.
}
\description{
Pull the final items of all players in the match.
}
\examples{
\dontrun{
a <- get_match('6156757097')
b <- pull_player_final_items(a)
}
}
